Move some functions to pidvelmap module
