+++ src/mightybug_a/main.h
@@ -15,3 +15,4 @@
 #include "telemetry.h"
 #include "dma.h"
 #include "mapping.h"
+#include "pidvelmap.h"
